core-components: split out LogViewerControls
